Revert "[mlir][affine] allow iter args as valid dims" (#140808)
Reverts llvm/llvm-project#139069: the original change that got reverted
by that PR was in fact correct. Since we don't know how secondary
iteration args evolve with loop iterations (e.g., the loop may be
yielding the result of a function call or the state of a PRNG), we
cannot statically represent them as an integer set bound by affine
constraints, unlike primary iteration args where the evolution is clear.