[ELF,test] Make loongarch-reloc-leb128.s similar to riscv-reloc-leb128.s
For the label difference A-B where A and B are in the same section,
if the section contains no linker-relaxable instruction, we can disable
the framnent walk code path. However, the test from #81133 relies on the
redundant relocations. Make it similar to riscv-reloc-leb128.s.
For now, lose the 32-bit coverage as call36 is 64-bit only.